refactor(table): extract index metadata maintenance into its own module
An index entry claims coverage of a set of fragments and fields, and any
operation that rewrites data can invalidate part of that claim. The ten methods
that narrow a fragment bitmap, drop no-longer-described fields, or drop the
index outright were spread through the second half of impl Transaction.

Getting these rules wrong does not fail a commit, it silently returns stale rows
from the index, so they are worth reading as a group next to the 20 tests that
pin them down.
